Added barrelling for media controls bar
